Updated documentation

+++ source/graphics/graphics.cpp
@@ -71,7 +71,7 @@ Graphics::generateText(std::string text, const int& size) {
 
     int       pos         = 0;
     SDL_FRect selector    = { 0, 0, 8.0f, 8.0f };
-    SDL_FRect destination = { 0, 0, size, size };
+    SDL_FRect destination = { 0, 0, static_cast<float>(size), static_cast<float>(size) };
     for (const auto& c : text) {
         destination.x = static_cast<float>(pos++) * size;
         if ((int)c > 0x40 && (int)c < 0x91) {
